pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI boilerplate"
authors = ["dmytro <dvbarchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
alembic = "^1.14.0"
logfire = {extras = ["fastapi"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.9.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
